Move HomePage to its own folder in `components/pages/`
