+++ components/pages/HomePage/AuthorView.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { Text, View, Platform } from "react-native";
-import { FONTS } from "../../helpers/constants";
-import { Author } from "../../helpers/wpapi";
+import { FONTS } from "../../../helpers/constants";
+import { Author } from "../../../helpers/wpapi";
 
 export const AuthorView: React.ElementType = ({
   authors,
