+++ components/pages/HomePage/GrindSection.tsx
