+++ components/pages/HomePage/MainSection.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { MediaRule, mergeRStyle } from "emotion-native-media-query";
-import { BREAKPOINTS } from "../../helpers/constants";
-import { Section, SECTION_PADDING } from "../Section";
+import { BREAKPOINTS } from "../../../helpers/constants";
+import { Section, SECTION_PADDING } from "../../Section";
 import { HeadlineArticle } from "./HeadlineArticle";
 import { TopThumbnailArticle } from "./TopThumbnailArticle";
 import { Column } from "./Column";
