+++ components/pages/HomePage/MoreFromTheDailySection.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Text, TouchableOpacity } from "react-native";
 import RView from "emotion-native-media-query";
-import { STANFORD_COLORS } from "../../helpers/constants";
-import { Post } from "../../helpers/wpapi";
-import { Section, SectionStyle, SECTION_PADDING } from "../Section";
+import { STANFORD_COLORS } from "../../../helpers/constants";
+import { Post } from "../../../helpers/wpapi";
+import { Section, SectionStyle, SECTION_PADDING } from "../../Section";
 import { TextOnlyArticle } from "./TextOnlyArticle";
 import { SectionTitle } from "./SectionTitle";
 import { SectionProps } from "./SectionProps";
