+++ components/pages/HomePage/OpinionSection.tsx
