+++ components/pages/HomePage/PostExcerpt.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Text } from "react-native";
-import { FONTS } from "../../helpers/constants";
+import { FONTS } from "../../../helpers/constants";
 import { ArticleProps } from "./ArticleProps";
 
 export const PostExcerpt: React.ElementType = ({
