+++ components/pages/HomePage/SectionProps.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Style, RStyle } from "emotion-native-media-query";
-import { Post } from "../../helpers/wpapi";
+import { Post } from "../../../helpers/wpapi";
 // TODO: layout got reset to mobile one when returning from other app on iPad
 export interface SectionProps {
   content: Post[];
