+++ components/pages/HomePage/SponsoredSection.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Section } from "../Section";
+import { Section } from "../../Section";
 import { SectionTitle } from "./SectionTitle";
 import { SectionProps } from "./SectionProps";
