+++ components/pages/HomePage/TitleOnlyArticle.tsx
@@ -1,22 +1,13 @@
 import React from "react";
-import { Article, ArticleHeader } from "../Article";
+import { Article, ArticleHeader } from "../../Article";
 import { AuthorView } from "./AuthorView";
 import { ArticleProps } from "./ArticleProps";
 import { ArticleTitleWithLink } from "./ArticleTitleWithLink";
-import { ThumbnailImageWithLink } from "./ThumbnailImageWithLink";
 
-export const TopThumbnailArticle: React.ElementType = ({
-  post,
-}: ArticleProps) => {
+export const TitleOnlyArticle: React.ElementType = ({ post }: ArticleProps) => {
   const { tsdAuthors } = post;
   return (
     <Article post={post}>
-      <ThumbnailImageWithLink
-        post={post}
-        style={{
-          height: 120,
-        }}
-      />
       <ArticleHeader>
         <ArticleTitleWithLink post={post} />
       </ArticleHeader>
