+++ components/pages/HomePage/TopSection.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Text, View } from "react-native";
 import RView, { MediaRule } from "emotion-native-media-query";
-import { BREAKPOINTS, STANFORD_COLORS } from "../../helpers/constants";
-import { SectionStyle } from "../Section";
-import { OrderedList } from "../List";
+import { BREAKPOINTS, STANFORD_COLORS } from "../../../helpers/constants";
+import { SectionStyle } from "../../Section";
+import { OrderedList } from "../../List";
 import { getBorderValue } from "./getBorderValue";
 import { SectionProps } from "./SectionProps";
 import { LinkToArticle } from "./LinkToArticle";
